Select equipped item with number keys now

+++ net.h
@@ -65,7 +65,7 @@ typedef struct
 typedef struct
 {
     double delta_t;
-    uint16_t keys;
+    uint32_t keys;
     int mouse_x;
     int mouse_y;
 } __attribute__((__packed__)) NetPlayerInput;
